Create perceptron_movie_decision.py
Title: Perceptron Networks Aim:Implement a python program for Perceptron Networks by considering the given scenario. A student wants to make a decision about whether to go for a movie or not by looking at 3 parameters using a single neuron. The three inputs are Favorite hero, Exam, and Climate. Each has weights and we have a bias in the perceptron. If the condition is true input is 1 else input is 0, weights for Favorite hero=0.2, Exam=0.4, and Climate=0.2 and bias=-0.5. Output is 1. The decision is to go for a movie.Calculate the Accuracy Tools: None Procedure:

Initialize the input vector.
Train the network weights for the perceptron. DEEP LEARNING LABORATORY –AIL66/AD
Make predictions with the perceptron. ● import the necessary libraries ●Assign the input features to x ● Assign the target features to y ● Initialize the Perceptron with the appropriate number of inputs ● Train the model ● Predict from the test dataset ● Find the accuracy of the model